✨ APIサービスを統合し、ユーザー取得機能を更新
